requirements.txt
# requirements.txt - исправленные версии
aiogram==3.4.1
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
celery==5.3.4
fastapi==0.108.0
uvicorn==0.25.0
python-dotenv==1.0.0
aiohttp==3.9.1
feedparser==6.0.10
openai==1.7.0
cryptography>=42.0.0,<46.0.0
psycopg2-binary==2.9.9